% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toad_example.r
\name{hello_world}
\alias{hello_world}
\title{printf Redirected to Rprinft}
\usage{
hello_world()
}
\description{
This function provides a proof of concept of a C redirect
}
\examples{
# Gregor's suggestion
captured_data = capture.output(hello_world())

# Using sink around multiple function calls to redirect output
# to a single file
sink("sink-examp.txt")
hello_world()
sink()

input_data = readLines("sink-examp.txt")

all.equal(input_data, captured_data)
}
