% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toad_example.r
\docType{package}
\name{printf2Rprintf-package}
\alias{printf2Rprintf}
\alias{printf2Rprintf-package}
\title{printf2Rprintf: Automatic conversion of printf to Rprintf}
\description{
Shows how to bind printf to Rprintf and then capture the resulting
output.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
