% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printf2Rprintf-package.R
\docType{package}
\name{printf2Rprintf-package}
\alias{printf2Rprintf}
\alias{printf2Rprintf-package}
\title{printf2Rprintf: Automatic conversion of printf to Rprintf}
\description{
Shows how to bind the C function printf() to R's Rprintf() and then capture the resulting output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless-rd-rcpp/r-c-api-remap-printf}
  \item \url{http://rd-rcpp.thecoatlessprofessor.com/r-c-api-remap-printf/}
  \item Report bugs at \url{https://github.com/coatless-rd-rcpp/r-c-api-remap-printf}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
